Refactor 'otf' CLI
